Update MovieServiceImpl
Bug where trying to add a new movie instead replaces movie with id = 1. 
When navigating to /movie/add, the findById() service is called with null ID. This poplates the form with id reference = 1.
When this form is sent via post method, an ID of 1 is passed. 
Changed the default movie else block to include all fields except ID for the user form. When sent via post method, this will pass an empty ID field which will be auto incremented during the repo.saveAndFlush() method.